+++ shell.h
@@ -8,9 +8,9 @@ class ShellSort : public Sort {
         ShellSort(int *elements, size_t size) : Sort(elements, size) {}
 
         void execute() override {
-            // TODO
-            for (int gap = size/2; gap > 0; gap /= 2) {
-                for (int j = gap; j < size; ++j) {
+            // TO DO
+            for (int gap = static_cast<int>(size) / 2; gap > 0; gap /= 2) {
+                for (int j = gap; j < static_cast<int>(size); ++j) {
                     int k = j;
                     while (k >= gap && array[k - gap] > array[k]) {
                         std::swap(array[k - gap], array[k]);
